+++ plugin/app/Models/GradingMethod.php
@@ -30,7 +30,7 @@ class GradingMethod extends Model
         return $this->code === 2;
     }
 
-    public function isPreferBestEachTestGrade()
+    public function isPreferBestEachGrade()
     {
         return $this->code === 3;
     }